Finalise serverless demo
