--- aws_topology/tests/cloudformation_infra/lambdas/notification-send/Pipfile
@@ -1,16 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-pylint = "*"
-awscli = "*"
-autopep8 = "*"
-
-[packages]
-boto3 = "*"
-psycopg2-binary = "*"
-
-[requires]
-python_version = "3.8"
